sign 90%

+++ snsproject/app/src/main/java/com/example/sns_project/SignActivity.java
@@ -10,15 +10,36 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.android.volley.DefaultRetryPolicy;
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.JsonObjectRequest;
+import com.android.volley.toolbox.Volley;
 import com.example.sns_project.databinding.ActivitySignUpBinding;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.SetOptions;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.util.HashMap;
+import java.util.Map;
 
 public class SignActivity extends AppCompatActivity {
+    private String BUSINESSNUMBER = "";
+    final String[] ANSWER = new String[1];
+    private boolean businessNumCheck;
+    private long backKeyPressedTime = 0;
     private FirebaseAuth mAuth;
+    private UserProfile userProfile = new UserProfile();
+    FirebaseFirestore db = FirebaseFirestore.getInstance();
     String location;
     private ActivitySignUpBinding binding;
 
@@ -34,6 +55,26 @@ public class SignActivity extends AppCompatActivity {
         adapter.setDropDownViewResource(R.layout.spinner_item);
         set_SpinnerAdapter(adapter,arr);
 
+        binding.BNCheckBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                String businessNum = binding.businessNum.getText().toString();
+
+                if(businessNum.length() == 10) {
+                    request(businessNum);
+                    if (ANSWER[0] != null) {
+                        if (BN_CheckString(ANSWER[0])) {
+                            businessNumCheck = true;
+                            BUSINESSNUMBER = businessNum;
+                        }
+                    }
+                }
+                else{
+                    Tost("사업자등록번호를 입력해주세요.");
+                }
+            }
+        });
+
         binding.signupBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -47,6 +88,11 @@ public class SignActivity extends AppCompatActivity {
 
                 if(!binding.checkBoxAgree.isChecked()){
                     Tost("동의하기 버튼에 체크해주세요.");
+                    return;
+                }
+
+                if(!businessNumCheck){
+                    Tost("사업자번호를 확인버튼을 눌러 검증해주세요.");
                     return;
                 }
 
@@ -78,6 +124,7 @@ public class SignActivity extends AppCompatActivity {
                         if (task.isSuccessful()) {
                             Tost("가입되었습니다!");
                             FirebaseUser user = mAuth.getCurrentUser();
+                            BusinessNumberUpdate(user);
                             MainActivity();
                         } else {
                            Tost("이메일형식/비밀번호를 확인해주세요.");
@@ -87,8 +134,20 @@ public class SignActivity extends AppCompatActivity {
                 });
     }
 
-    public boolean inputcheck(String m, String p, String pc, String nickname, String phone, String store, String location){
-        boolean[] flag = {false,false,false,false,false,false};
+    public void BusinessNumberUpdate(FirebaseUser user){
+
+        Map<String, Object> data = new HashMap<>();
+        data.put(userProfile.ID, user.getUid());
+        data.put(userProfile.location, location);
+        data.put(userProfile.phone, binding.phoneNum.getText().toString());
+        data.put(userProfile.store, binding.storeName.getText().toString());
+        data.put(userProfile.BusinessNumber, BUSINESSNUMBER);
+        db.collection(userProfile.USER).document(user.getUid()).set(data, SetOptions.merge());
+
+    }
+
+    public boolean inputcheck(String m, String p, String pc, String nickname, String phone, String store,String location){
+        boolean[] flag = {false,false,false,false,false,false,false};
         if(nickname.length() >= 1 && m.length() >= 1)
             flag[0] = true;
         if(p.length() >= 6)
@@ -138,6 +197,62 @@ public class SignActivity extends AppCompatActivity {
         return true;
     }
 
+    public void request(String str){
+
+        //JSON형식으로 데이터 통신
+        String url = "https://jinlab.herokuapp.com/";
+        JSONObject testjson = new JSONObject();
+
+        try {
+            //입력해둔 edittext의 id와 pw값을 받아와 put해줍니다 : 데이터를 json형식으로 바꿔 넣어주었습니다.
+            testjson.put("id",str);
+            String jsonString = testjson.toString(); //완성된 json 포맷
+
+            //이제 전송
+            final RequestQueue requestQueue = Volley.newRequestQueue(SignActivity.this);
+            final JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.POST, url,testjson, new Response.Listener<JSONObject>() {
+
+                //데이터 전달을 끝내고 이제 그 응답을 받을 차례입니다.
+                @Override
+                public void onResponse(JSONObject response) {
+                    try {
+                        //받은 json형식의 응답을 받아
+                        JSONObject jsonObject = new JSONObject(response.toString());
+                        String res = jsonObject.getString("approve_id");
+
+                        if(res != null){
+                            ANSWER[0] = res;
+                            Tost(res);
+                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                }
+                //서버로 데이터 전달 및 응답 받기에 실패한 경우 아래 코드가 실행됩니다.
+            }, new Response.ErrorListener() {
+                @Override
+                public void onErrorResponse(VolleyError error) {
+                    error.printStackTrace();
+                    //Toast.makeText(MainActivity.this, error.toString(), Toast.LENGTH_SHORT).show();
+                }
+            });
+            jsonObjectRequest.setRetryPolicy(new DefaultRetryPolicy(DefaultRetryPolicy.DEFAULT_TIMEOUT_MS, DefaultRetryPolicy.DEFAULT_MAX_RETRIES, DefaultRetryPolicy.DEFAULT_BACKOFF_MULT));
+            requestQueue.add(jsonObjectRequest);
+            //
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public boolean BN_CheckString(String res){
+
+        if(res.length() == 16){
+            return true;
+        }else
+            return false;
+
+    }
+
     public void Tost(String str){
         Toast.makeText(this,str,Toast.LENGTH_SHORT).show();
     }
@@ -155,10 +270,19 @@ public class SignActivity extends AppCompatActivity {
 
     @Override
     public void onBackPressed() {
-        super.onBackPressed();
-        moveTaskToBack(true);
-        android.os.Process.killProcess(android.os.Process.myPid());
-        System.exit(1);
+        // 마지막으로 뒤로가기 버튼을 눌렀던 시간 저장
+        if (System.currentTimeMillis() > backKeyPressedTime + 1500) {
+            backKeyPressedTime = System.currentTimeMillis();
+          Tost("\'뒤로\' 버튼을 한번 더 누르시면 종료됩니다.");
+            return;
+        }
+        if (System.currentTimeMillis() <= backKeyPressedTime + 1500) {
+            //아래 3줄은 프로세스 종료
+            moveTaskToBack(true);
+            android.os.Process.killProcess(android.os.Process.myPid());
+            System.exit(1);
+        }
+
     }
 
     public void set_SpinnerAdapter( ArrayAdapter adapter,String[] arr ){
